vpr: Use a binary-tree reduction to incrementally update placer timing costs
This allows the placer timing costs to be updated in O(klogK) time, where 'K'
is the total number of connections in the netlist, and 'k' is the number of
modified connections since the last re-calculation.

This is done with an implicit binary-tree to ensure the same order of
floating point operations and avoid numeric instability.